[List] Add keyboard navigation
